remove copies, rename currently work on generator as main file
